Fix bug in dc code

+++ src/search_algorithms/search.h
@@ -11,5 +11,5 @@ int generate_next_partition(int* a, int* b, int n);
 
 // Helper functions for divide and conquer
 int division(int move_from, int move_to, mcm& model);
-int find_member_i(__uint128_t community, int i);
-int index_of_member(int member);
+__uint128_t find_member_i(__uint128_t community, int i);
+int index_of_member(__uint128_t member);
